refactor: reorganize ingestion modules into a Legacy directory and relocate utility and utility-adjacent files.
