+++ src/ravioli/backend/data/olap/ingestion/Legacy/apple_health.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from tqdm import tqdm
 from psycopg2 import sql
 from psycopg2.extras import execute_values
-from ravioli.backend.data.olap.ingestion.base import BaseIngestor
+from ravioli.backend.data.olap.ingestion.Legacy.base import BaseIngestor
 from ravioli.backend.data.oltp.session import get_db_connection
 from ravioli.backend.core.config import settings
